# contracts/openapi/dokpod-control-plane.v1.yaml
openapi: 3.1.0
info:
  title: Dokpod Control Plane API
  version: 1.0.0
  description: Contrato público inicial do cadastro de ambientes do Dokpod.
servers:
  - url: https://localhost:7443
paths:
  /api/v1/environments/{environmentId}/commands/{commandId}:
    get:
      operationId: getContainerCommand
      summary: Consulta o estado de um comando de ciclo de vida
      security:
        - bearerAuth: []
      parameters:
        - name: environmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: commandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Estado atual autorizado do comando.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerCommandStatus'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/v1/environments/{environmentId}/containers/{containerId}/commands:
    post:
      operationId: submitContainerCommand
      summary: Solicita uma ação de ciclo de vida em um container
      security:
        - bearerAuth: []
      parameters:
        - name: environmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: containerId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{64}$'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: Identifica de forma estável a intenção lógica para retries seguros.
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitContainerCommandRequest'
      responses:
        '202':
          description: Comando persistido ou replay idempotente aceito para processamento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerCommandAccepted'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Identificador técnico opcional para correlação server-side.
  schemas:
    SubmitContainerCommandRequest:
      type: object
      required: [action, expectedContainerRevision, deadlineUtc]
      additionalProperties: false
      properties:
        action:
          type: string
          enum: [start, stop, restart, delete]
        expectedContainerRevision:
          type: string
          minLength: 1
          maxLength: 255
        deadlineUtc:
          type: string
          format: date-time
    ContainerCommandAccepted:
      type: object
      required: [commandId]
      additionalProperties: false
      properties:
        commandId:
          type: string
          format: uuid
    ContainerCommandStatus:
      type: object
      required:
        - environmentId
        - commandId
        - action
        - containerId
        - expectedContainerRevision
        - state
        - failureCode
        - observedContainerRevision
        - deadlineUtc
        - createdAtUtc
        - updatedAtUtc
        - completedAtUtc
      additionalProperties: false
      properties:
        environmentId:
          type: string
          format: uuid
        commandId:
          type: string
          format: uuid
        action:
          type: string
          enum: [start, stop, restart, delete]
        containerId:
          type: string
          pattern: '^[0-9a-fA-F]{64}$'
        expectedContainerRevision:
          type: string
          minLength: 1
          maxLength: 255
        state:
          type: string
          enum: [pending, dispatched, accepted, succeeded, failed, indeterminate]
        failureCode:
          type: [string, 'null']
          maxLength: 128
        observedContainerRevision:
          type: [string, 'null']
          maxLength: 255
        deadlineUtc:
          type: string
          format: date-time
        createdAtUtc:
          type: string
          format: date-time
        updatedAtUtc:
          type: string
          format: date-time
        completedAtUtc:
          type: [string, 'null']
          format: date-time
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        traceId:
          type: string
        code:
          type: string
  responses:
    InvalidRequest:
      description: Requisição inválida.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    AuthenticationRequired:
      description: Autenticação ausente ou inválida.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    AccessDenied:
      description: Decisão do Keycloak negou o cadastro.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: O ambiente já existe.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: O ambiente não foi encontrado após autorização.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    DependencyUnavailable:
      description: Autorização ou persistência indisponível.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'